refactor assets add files image&svg&videos
