+++ src/assets/images/sentIcon.png
