+++ src/pages/DoctorProfile/EditProfile.jsx
@@ -4,14 +4,14 @@ import { useEffect, useMemo, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import styled from "styled-components";
 import { create } from "zustand";
-import AuthMenImg from "../../assets/authMen.png";
-import AuthWomenImg from "../../assets/authWomen.png";
-import EditImage from "../../assets/editImage.svg";
-import EditProfileIcon from "../../assets/editProfileIcon.svg";
-import ProfileDoctor from "../../assets/profileDoctor.svg";
-import SaveProfile from "../../assets/save.svg";
-import CancelIcon from "../../assets/x-square.svg";
-import RemoveIcon from "../../assets/x.svg";
+import AuthMenImg from "../../assets/images/authMen.png";
+import AuthWomenImg from "../../assets/images/authWomen.png";
+import ProfileDoctor from "../../assets/svg/doctorProfileIcon.svg";
+import EditProfileIcon from "../../assets/svg/editIcon.svg";
+import EditImage from "../../assets/svg/editImgIcon.svg";
+import SaveProfile from "../../assets/svg/saveIcon.svg";
+import CancelIcon from "../../assets/svg/x-square.svg";
+import RemoveIcon from "../../assets/svg/xIcon.svg";
 import { PopUp } from "../../components/PopUp";
 import { fireDb, fireStorage } from "../../firebase";
