+++ src/pages/Registration/index.jsx
@@ -4,9 +4,9 @@ import { useForm } from "react-hook-form";
 import { useLocation, useNavigate } from "react-router-dom";
 import styled, { keyframes } from "styled-components";
 import { fetchDepartmentsData } from "../../api";
-import AnnouncementImg from "../../assets/announcementImage.svg";
-import BackGround from "../../assets/background.svg";
-import BackGroundMobile from "../../assets/background_mobile.svg";
+import AnnouncementImg from "../../assets/svg/announcementImage.svg";
+import BackGround from "../../assets/svg/background.svg";
+import BackGroundMobile from "../../assets/svg/backgroundMobile.svg";
 import AiRobot from "../../components/AiRobot";
 import CusService from "../../components/CusService";
 import Loading from "../../components/Loading";
